backend/requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Spotify API Client
spotipy==2.23.0

# Database
psycopg2-binary==2.9.9

# Data Processing & ML
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
